[dataprotect] BM-16457: feat: use pg_dump format custom instead of format tar.
The custom format is more optimized in many ways: data is compressed by default,
and this format does not requires temporary disk usage on /tmp when doing pg_dump.

The format is compatible with pg_restore, so no modifications are needed for the restoration.
Old format is compatible with dataprotect restore without any modification.