bugfix: installation somestimes fails
In some cases, vert.x chooses to undeploy a verticle because of a failure during start().
This failure causes vert.x to close the workerPool associated with it.
When this happens, bluemind is unable to run anything from rest services, with this error:

net.bluemind.core.api.fault.ServerFault: Worker executor closed