BM-17470 [MailApp] [conversation] Fix: make conversation.messages property an array of references to messages
conversation.messages was a an array of object duplicating information
from store.conversations.messages. To keep store consitency and to be
more compliant with the store philosophy this property is now a array of
references to store.conversations.messages.
The consistency between is only managed throw hook on ADD / REMOVE messages
and on ADD / REMOVE conversations.
The consistency of selection / conversationList is now also only managed
throw hook on conversation mutation.