BM-18100 fix: pg_repack was in conflict with pg_dump
pg_repack needs an exclusive lock for a short amount of time on each tables.
If pg_repack can't get the lock after (by default 60 seconds) 4 hours, the offending backend is killed.

Also, move the Maintenance job to 1 AM instead of 3 AM to avoid conflicting with pg_dump as much as possible.