[mapi] BM-18204 fix: invalidate mapdb immediately after removal
The removalListener of caffeine cache can run when it wants when the cause is "Eviction", not synchronously.
We need to close the mapdb as soon as possible for a given mapiContext, in order
to avoid a race with a newly created mapi connection/context.

evictionListener on the other hand, will be called during the atomic removal, which
is what we need in this case.