BM-18869 fix: remake the bm-postgresql-postupgrade asynchronous: avoid blocking upgrades
We used to use a bm-postgresql-postupgrade service Type=simple, which forks a process, making
the postupgrade step fully asynchronous.

However, it was changed to Type=oneshot, which is fully synchronous, thus, blocking the bm-postgresql service
during an upgrade, which is the oposite of what we want.

This patch makes the postupgrade.sh run as a background process using bash's `(subshell) & disown`