[node] feat: make writeFile "atomic", by using a temporary file while writing.
Uses a temporary file for writing the data, in the same directory to avoid cross filesystem copies.
If the file is successfully written, then and only then, the file is moved to it's final name.

This avoids a case when multiple clients try to write to the same file, or
when someone tries to upload something, but, while uploading, another one tries to read it. (In such cases, the data is not yet fully written)