[imap,pop,*netserver] fix: Finally found what was wrong about NetSocket / vertx context
So, NetSocket created by NetServer is tied to the context used by listen().
Usually, it's an eventloop root context.

If we want to use our own DuplicatedContext(), we need to ensure *NOT*
to use any Future created by the NetSocket directly, and force all
AsyncResult to be produced by the correct context.

So, we just wrap the NetSocket right from the begenning and we are fine!