chore: fixing AppendTests
In the multithreaded tests, every test tries to create an email, marks it deleted, then expunges the whole lot at the end.
The test reuses the deleted mail as an append, but when an email is really
removed by expunge by another thread, the fetch may return an empty response.

We must avoid using the message in this case.